# api.yaml
openapi: 3.0.0
info:
  title: Event Management API
  version: 1.0.0
  description: API for managing organizations, events, and transactions.
servers:
  - url: https://api.example.com/v1
    description: Production server
security:
  - bearerAuth: []
paths:
...
  /organizations/{orgId}/events:
    get:
      summary: Get all events for a specific organization
      operationId: getEventsForOrganization
      security:
        - bearerAuth: []
      parameters:
        - name: orgId
          in: path
          required: true
          description: ID of the organization
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: A list of events for the organization
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Event'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
    post:
      summary: Create a new event for an organization
      operationId: createEventForOrganization
      security:
        - bearerAuth: []
      parameters:
        - name: orgId
          in: path
          required: true
          description: ID of the organization
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewEvent'
      responses:
        '201':
          description: Event created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /organizations/{orgId}/events/{eventId}:
    get:
      summary: Get a specific event by ID for an organization
      operationId: getEventByIdForOrganization
      security:
        - bearerAuth: []
      parameters:
        - name: orgId
          in: path
          required: true
          description: ID of the organization
          schema:
            type: string
            format: uuid
        - name: eventId
          in: path
          required: true
          description: ID of the event to retrieve
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Event details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
    put:
      summary: Update an existing event for an organization
      operationId: updateEventForOrganization
      security:
        - bearerAuth: []
      parameters:
        - name: orgId
          in: path
          required: true
          description: ID of the organization
          schema:
            type: string
            format: uuid
        - name: eventId
          in: path
          required: true
          description: ID of the event to update
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewEvent'
      responses:
        '200':
          description: Event updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
    delete:
      summary: Delete an event for an organization
      operationId: deleteEventForOrganization
      security:
        - bearerAuth: []
      parameters:
        - name: orgId
          in: path
          required: true
          description: ID of the organization
          schema:
            type: string
            format: uuid
        - name: eventId
          in: path
          required: true
          description: ID of the event to delete
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Event deleted successfully
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Event:
      type: object
      required:
        - id
        - name
        - date
        - venue
        - entryFee
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier for the event
          readOnly: true
        name:
          type: string
          description: Name of the event
        date:
          type: string
          format: date
          description: Date of the event (YYYY-MM-DD)
        venue:
          type: string
          description: Venue of the event
        entryFee:
          type: number
          format: float
          description: Entry fee for the event
          minimum: 0
        metadata:
          type: object
          description: Additional metadata for the event
          nullable: true
          properties:
            description:
              type: string
            capacity:
              type: integer
              minimum: 0
    NewEvent:
      type: object
      required:
        - name
        - date
        - venue
        - entryFee
      properties:
        name:
          type: string
          description: Name of the event
        date:
          type: string
          format: date
          description: Date of the event (YYYY-MM-DD)
        venue:
          type: string
          description: Venue of the event
        entryFee:
          type: number
          format: float
          description: Entry fee for the event
          minimum: 0
        metadata:
          type: object
          description: Additional metadata for the event
          nullable: true
          properties:
            description:
              type: string
            capacity:
              type: integer
              minimum: 0
...
  responses:
    UnauthorizedError:
      description: Authentication information is missing or invalid.
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: Unauthorized
    NotFoundError:
      description: The specified resource was not found.
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: Resource not found
    BadRequest:
      description: Invalid request payload or parameters.
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: Invalid input